[SQUASH ME SOMEWHERE] Update huffman related datatype.
Important to note the __attribute__((__packed__)) that avoid
misalignment by compiler of the struct fields (so we can read from
memory directly without problems).

+++ huffman/huffman.c
@@ -14,7 +14,7 @@
 
 
 huffman_t* Huffman = NULL;
-unsigned int HuffmanLength;
+uint8_t HuffmanLength;
 unsigned int HuffmanIdx = 0;
 
 static int cmp_nodes(const void* a, const void* b) {
